aula 3 - consulta com jpql @query + parametros

+++ src/main/java/br/com/alura/springdata/repository/FuncionarioRepository.java
@@ -1,5 +1,6 @@
 package br.com.alura.springdata.repository;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -13,12 +14,9 @@ import br.com.alura.springdata.orm.Funcionario;
 public interface FuncionarioRepository extends JpaRepository<Funcionario, Long>{
 	List<Funcionario> findByNome(String nome);
 	
-	// método funcional porém desnecessáriamente verboso
-	//List<Funcionario> findByNomeAndSalarioGreaterThanAndDataContratacao(String nome, Double salario, LocalDate data);
-	
 	@Query("SELECT f FROM Funcionario f "
 			+ "WHERE f.nome = :nome "
 			+ "AND f.salario >= :salario "
 			+ "AND f.dataContratacao = :data ")
-	List<Funcionario> findNomeSalarioMaiorDataContratacao(String nome, Double salario, LocalDate data);
+	List<Funcionario> findNomeSalarioMaiorDataContratacao(String nome, BigDecimal salario, LocalDate data);
 }
